+++ src/components/board/BoardDetail/CompanyModal.js
@@ -1,65 +1,63 @@
-// import React, { useState } from 'react';
-import { useEffect, useState } from "react";
-import { useParams } from "react-router-dom";
-import { Button, Modal, Card } from 'react-bootstrap';
-import { createBrowserHistory } from 'history'; // 모달주소이동
+// // import React, { useState } from 'react';
+// import { useEffect, useState } from "react";
+// import { useParams } from "react-router-dom";
+// import { Button, Modal, Card } from 'react-bootstrap';
+// import { createBrowserHistory } from 'history'; // 모달주소이동
 
-// 아이콘
-import { FaPlane } from "react-icons/fa6";
-import { MdOutlineLocationOn, MdDateRange } from "react-icons/md";
-import { IoPersonCircleOutline } from "react-icons/io5";
+// // 아이콘
+// import { FaPlane } from "react-icons/fa6";
+// import { MdOutlineLocationOn, MdDateRange } from "react-icons/md";
+// import { IoPersonCircleOutline } from "react-icons/io5";
 
-// toast
-import { ToastContainer, toast } from 'react-toastify';
+// // toast
+// import { ToastContainer, toast } from 'react-toastify';
 
-// 시간
-import moment from 'moment';
+// // 시간
+// import moment from 'moment';
 
-import { getCompanyBaordList } from '../../../api/BoardApi';
-import { postReply } from "../../../api/BoardApi";
-import { getReply } from "../../../api/BoardApi";
+// import { getCompanyBaordList } from '../../../api/BoardApi';
+// import { postReply } from "../../../api/BoardApi";
+// import { getReply } from "../../../api/BoardApi";
 
-// css
-import 'react-toastify/dist/ReactToastify.css';
-import '../../../styles/board/companyModal.scss';
-
-
-function CompanyModal() {
-
-  const [show, setShow] = useState(false);
-  const [selectedItem, setSelectedItem] = useState(null);
-  const history = createBrowserHistory();
-
-  const handleShow = (bno) => {
-    setShow(true);
-    setSelectedItem(bno);
-    history.push(`/board/company/${bno}`);
-  };
-
-  const handleClose = () => {
-    setShow(false);
-    setSelectedItem(null);
-    history.push('/board/company');
-  };
-
-  const [commentContent, setCommentContent] = useState('');
-
-  // const handleClick = () => { // toast
-  //   toast.info('신청되었습니다!', {
-  //     position: "bottom-center",
-  //     autoClose: 1000,
-  //     hideProgressBar: true, // 프로그래스바 안 보이도록
-  //     newestOnTop: false,
-  //     closeOnClick: true,
-  //     rtl: false,
-  //     pauseOnFocusLoss: true,
-  //     draggable: true,
-  //     pauseOnHover: true,
-  //     theme: 'light'
-  //   });
-  // };
+// // css
+// import 'react-toastify/dist/ReactToastify.css';
+// import '../../../styles/board/companyModal.scss';
 
 
+// function CompanyModal() {
+
+//   const [show, setShow] = useState(false);
+//   const [selectedItem, setSelectedItem] = useState(null);
+//   const history = createBrowserHistory();
+
+//   const handleShow = (bno) => {
+//     setShow(true);
+//     setSelectedItem(bno);
+//     history.push(`/board/company/${bno}`);
+//   };
+
+//   const handleClose = () => {
+//     setShow(false);
+//     setSelectedItem(null);
+//     history.push('/board/company');
+//   };
+
+//   const [commentContent, setCommentContent] = useState('');
+
+//   // const handleClick = () => { // toast
+//   //   toast.info('신청되었습니다!', {
+//   //     position: "bottom-center",
+//   //     autoClose: 1000,
+//   //     hideProgressBar: true, // 프로그래스바 안 보이도록
+//   //     newestOnTop: false,
+//   //     closeOnClick: true,
+//   //     rtl: false,
+//   //     pauseOnFocusLoss: true,
+//   //     draggable: true,
+//   //     pauseOnHover: true,
+//   //     theme: 'light'
+//   //   });
+//   // };
 
 
 
@@ -69,182 +67,184 @@ function CompanyModal() {
 
 
 
-  // 댓글 가져오기
-  const [replyList, setReplyList] = useState([]);
-  const { bno } = useParams();
-
-  useEffect(() => {
-    const fetchReplyData = async () => {
-      try {
-        const response = await getReply(); // 댓글을 가져오는 함수 호출
-        const data = response.data;
-        console.log("ReplyData: ", data);
-        setReplyList(data); // 반환된 데이터 배열을 상태로 설정 (댓글 목록)
-      } catch (error) {
-        console.error("Error fetching reply data:", error);
-      }
-    };
-
-    fetchReplyData(); // 댓글 가져오기 함수 호출
-  }, []);
 
 
+//   // 댓글 가져오기
+//   const [replyList, setReplyList] = useState([]);
+//   const { bno } = useParams();
 
+//   useEffect(() => {
+//     const fetchReplyData = async () => {
+//       try {
+//         const response = await getReply(); // 댓글을 가져오는 함수 호출
+//         const data = response.data;
+//         console.log("ReplyData: ", data);
+//         setReplyList(data); // 반환된 데이터 배열을 상태로 설정 (댓글 목록)
+//       } catch (error) {
+//         console.error("Error fetching reply data:", error);
+//       }
+//     };
 
-  // 여행메이트 게시판 글 가져오기
-  const [CompanyBoardListData, setCompanyBoardListData] = useState([]);
-
-  useEffect(() => {
-  const fetchReplyData = async () => {
-    try {
-      const response = await getReply(); // 댓글을 가져오는 함수 호출
-      const data = response.data;
-      console.log("응답 데이터:", data); // 응답 데이터를 로그에 남김
-      setReplyList(data); // 반환된 데이터 배열을 상태로 설정 (댓글 목록)
-    } catch (error) {
-      console.error("댓글 데이터를 불러오는 중 에러 발생:", error);
-    }
-  };
-
-  fetchReplyData(); // 댓글 가져오기 함수 호출
-}, []);
-
-
-  const handleSubmit = async (e, bno) => {
-    e.preventDefault();
-
-    try {
-      const commentData = {
-        content: commentContent,
-        bno: bno
-      };
-
-      const response = await postReply(commentData);
-      console.log(response);
-      alert('댓글 작성 완료');
-    } catch (error) {
-      console.log("댓글 작성 실패");
-      alert('댓글 작성 실패');
-    }
-  }
-
-  // textarea의 값을 상태로 업데이트하는 함수
-  const handleChange = (e) => {
-    setCommentContent(e.target.value);
-  }
-
-  return (
-    <>
-      <ToastContainer // toast 설정
-        position="bottom-center"
-        autoClose={1000}
-        hideProgressBar={false}
-        newestOnTop={false}
-        closeOnClick
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover
-        theme="light"
-      />
-
-
-      {CompanyBoardListData.map(item => (
-        <div key={item.bno} className='d-flex justify-content-center modal-'>
-          <Button className='apply-btn' onClick={() => handleShow(item.bno)}>
-            자세히
-            <FaPlane className='ariPlane-btn' />
-          </Button>
-          <Modal className='companyModal' show={show} onHide={handleClose}>
-            <Modal.Header className='modal-header' closeButton></Modal.Header>
-            <Modal.Body>
-              <div className="container">
-                <div className="row">
-                  <div className="area-left col-md-6">
-                    {/* 왼쪽 구역 */}
-                    <div className="info-left">
-                      <h2>{item.title}</h2>
-                      <div className="detail">
-                        <div>
-                          <p className='modal-title'><MdOutlineLocationOn className='location-icon' />{item.location}</p>
-                        </div>
-                        <div className='date-info'>
-                          <MdDateRange className='date-icon' />
-                          <Card.Text>{moment(item.regDate).format('YYYY/MM/DD')}</Card.Text>
-                        </div>
-                        <div >
-                          <Card.Text><IoPersonCircleOutline className="person-icno" />{item.id.id}</Card.Text>
-                        </div>
-                      </div>
-                      <div>
-                        {item.content}
-                      </div>
-                      <div className='Reply_div'>
-
-
-
-                        <div className='Reply_write'>
-                          <textarea
-                            rows='3'
-                            type="text"
-                            id="write_reply" // ID 변경
-                            placeholder='100자 이내의 글을 입력해주세요.'
-                            maxLength='100'
-                            name='write_reply'
-                            value={commentContent} // 상태와 연결
-                            onChange={handleChange} // 값이 변경될 때마다 상태 업데이트
-                          />
-                          <input type='submit' value='등록' id='reply_submit_button'
-                            onClick={(e) => handleSubmit(e, item.bno)} />
-                        </div>
-
-                        {/* <Form.Control
-                          rows='3'
-                          type="text"
-                          id="title"
-                          placeholder='100자 이내의 글을 입력해주세요.'
-                          maxLength='100'
-                          name='write_reply'
-                        onChange={handleChange}
-                        /> */}
-                        {/* <input type='submit' value='등록' id='reply_submit_button' onClick={handleSubmit}/> */}
-                        {/* <Button type="submit" id="reply_submit_button" onClick={handleSubmit}>
-                          등록
-                        </Button> */}
+//     fetchReplyData(); // 댓글 가져오기 함수 호출
+//   }, []);
 
 
 
 
+//   // 여행메이트 게시판 글 가져오기
+//   const [CompanyBoardListData, setCompanyBoardListData] = useState([]);
+
+//   useEffect(() => {
+//   const fetchReplyData = async () => {
+//     try {
+//       const response = await getReply(); // 댓글을 가져오는 함수 호출
+//       const data = response.data;
+//       console.log("응답 데이터:", data); // 응답 데이터를 로그에 남김
+//       setReplyList(data); // 반환된 데이터 배열을 상태로 설정 (댓글 목록)
+//     } catch (error) {
+//       console.error("댓글 데이터를 불러오는 중 에러 발생:", error);
+//     }
+//   };
+
+//   fetchReplyData(); // 댓글 가져오기 함수 호출
+// }, []);
 
 
-                      </div>
-                    </div>
-                  </div>
-                  {/* 오른쪽 구역 */}
-                  <div className="area-right col-md-6">
-                    <div className="Reply-list">
-                      {replyList.map((reply, index) => (
-                        <div key={index}>
-                          <div className="Reply-nav">
-                            <div className="nick-name">{reply.id}</div>
-                            <div className="time">{reply.regDate}</div>
-                            <Button className="delete" variant="link">삭제</Button>
-                          </div>
-                          <div>
-                            <div className="Reply-content">{reply.content}</div>
-                          </div>
-                        </div>
-                      ))}
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </Modal.Body>
-          </Modal>
+//   const handleSubmit = async (e, bno) => {
+//     e.preventDefault();
 
-        </div>
-      ))}
-    </>
-  );
-}
-export default CompanyModal;
+//     try {
+//       const commentData = {
+//         content: commentContent,
+//         bno: bno
+//       };
+
+//       const response = await postReply(commentData);
+//       console.log(response);
+//       alert('댓글 작성 완료');
+//     } catch (error) {
+//       console.log("댓글 작성 실패");
+//       alert('댓글 작성 실패');
+//     }
+//   }
+
+//   // textarea의 값을 상태로 업데이트하는 함수
+//   const handleChange = (e) => {
+//     setCommentContent(e.target.value);
+//   }
+
+//   return (
+//     <>
+//       <ToastContainer // toast 설정
+//         position="bottom-center"
+//         autoClose={1000}
+//         hideProgressBar={false}
+//         newestOnTop={false}
+//         closeOnClick
+//         rtl={false}
+//         pauseOnFocusLoss
+//         draggable
+//         pauseOnHover
+//         theme="light"
+//       />
+
+
+//       {CompanyBoardListData.map(item => (
+//         <div key={item.bno} className='d-flex justify-content-center modal-'>
+//           <Button className='apply-btn' onClick={() => handleShow(item.bno)}>
+//             자세히
+//             <FaPlane className='ariPlane-btn' />
+//           </Button>
+//           <Modal className='companyModal' show={show} onHide={handleClose}>
+//             <Modal.Header className='modal-header' closeButton></Modal.Header>
+//             <Modal.Body>
+//               <div className="container">
+//                 <div className="row">
+//                   <div className="area-left col-md-6">
+//                     {/* 왼쪽 구역 */}
+//                     <div className="info-left">
+//                       <h2>{item.title}</h2>
+//                       <div className="detail">
+//                         <div>
+//                           <p className='modal-title'><MdOutlineLocationOn className='location-icon' />{item.location}</p>
+//                         </div>
+//                         <div className='date-info'>
+//                           <MdDateRange className='date-icon' />
+//                           <Card.Text>{moment(item.regDate).format('YYYY/MM/DD')}</Card.Text>
+//                         </div>
+//                         <div >
+//                           <Card.Text><IoPersonCircleOutline className="person-icno" />{item.id.id}</Card.Text>
+//                         </div>
+//                       </div>
+//                       <div>
+//                         {item.content}
+//                       </div>
+//                       <div className='Reply_div'>
+
+
+
+//                         <div className='Reply_write'>
+//                           <textarea
+//                             rows='3'
+//                             type="text"
+//                             id="write_reply" // ID 변경
+//                             placeholder='100자 이내의 글을 입력해주세요.'
+//                             maxLength='100'
+//                             name='write_reply'
+//                             value={commentContent} // 상태와 연결
+//                             onChange={handleChange} // 값이 변경될 때마다 상태 업데이트
+//                           />
+//                           <input type='submit' value='등록' id='reply_submit_button'
+//                             onClick={(e) => handleSubmit(e, item.bno)} />
+//                         </div>
+
+//                         {/* <Form.Control
+//                           rows='3'
+//                           type="text"
+//                           id="title"
+//                           placeholder='100자 이내의 글을 입력해주세요.'
+//                           maxLength='100'
+//                           name='write_reply'
+//                         onChange={handleChange}
+//                         /> */}
+//                         {/* <input type='submit' value='등록' id='reply_submit_button' onClick={handleSubmit}/> */}
+//                         {/* <Button type="submit" id="reply_submit_button" onClick={handleSubmit}>
+//                           등록
+//                         </Button> */}
+
+
+
+
+
+
+//                       </div>
+//                     </div>
+//                   </div>
+//                   {/* 오른쪽 구역 */}
+//                   <div className="area-right col-md-6">
+//                     <div className="Reply-list">
+//                       {replyList.map((reply, index) => (
+//                         <div key={index}>
+//                           <div className="Reply-nav">
+//                             <div className="nick-name">{reply.id}</div>
+//                             <div className="time">{reply.regDate}</div>
+//                             <Button className="delete" variant="link">삭제</Button>
+//                           </div>
+//                           <div>
+//                             <div className="Reply-content">{reply.content}</div>
+//                           </div>
+//                         </div>
+//                       ))}
+//                     </div>
+//                   </div>
+//                 </div>
+//               </div>
+//             </Modal.Body>
+//           </Modal>
+
+//         </div>
+//       ))}
+//     </>
+//   );
+// }
+// export default CompanyModal;
